Move SlimeFiller to separate file. Move AnnotationConverter (fka SummaryHandler) to separate file.
