+++ searchsummary/src/vespa/searchsummary/docsummary/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchsummary_docsummary OBJECT
     SOURCES
+    annotation_converter.cpp
     array_attribute_combiner_dfw.cpp
     attribute_combiner_dfw.cpp
     attribute_field_writer.cpp
@@ -30,6 +31,7 @@ vespa_add_library(searchsummary_docsummary OBJECT
     resultconfig.cpp
     searchdatatype.cpp
     simple_dfw.cpp
+    slime_filler.cpp
     struct_fields_resolver.cpp
     struct_map_attribute_combiner_dfw.cpp
     summaryfeaturesdfw.cpp
